Separate new files by directories
