+++ SOGE/include/SOGE/Core/Event/EventCategory.hpp
@@ -1,5 +1,5 @@
-#ifndef SOGE_CORE_EVENTCATEGORY_HPP
-#define SOGE_CORE_EVENTCATEGORY_HPP
+#ifndef SOGE_CORE_EVENT_EVENTCATEGORY_HPP
+#define SOGE_CORE_EVENT_EVENTCATEGORY_HPP
 
 #include <variant>
 
@@ -81,4 +81,4 @@ namespace soge
     }
 }
 
-#endif // SOGE_CORE_EVENTCATEGORY_HPP
+#endif // SOGE_CORE_EVENT_EVENTCATEGORY_HPP
